added media queries to navbars
