+++ front/src/components/friend_request/FriendRequest.scss
@@ -8,6 +8,8 @@
 	align-items: center;
 	background-color: #fff;
 	border-radius: 6px;
+	background-color: white;
+	color: black;
 	box-shadow: 0px 0px 3px 0px rgb(22, 165, 58);
 	margin-bottom: 20px;
 	width: 500px;
@@ -75,6 +77,52 @@
 		&:hover div {
 			background-color: red;
 		}
+	}
+
+
+}
+@media (max-width: 500px) {
+	.friendRequestContainer {
+		width: 95%;
+		
+
+		&__title {
+			margin: 30px 0 0 0;
+			text-align: center;
+			font-size: 24px;
+		}
+
+		&__item {
+			display: block;
+			align-items: center;
+
+
+			&__avatar{
+				margin: 15px 70px;
+				width: 150px;
+				height: 150px;
+			}
+
+			&__username {
+				text-align: center;
+				font-size: 24px;
+			}
+
+			&__buttons {
+				margin: 15px 28px;
+
+				&--reject, &--accept{
+					font-size: 18px;
+				}
+			}	
+
+		}
+
+
+
+		
+
+
 	}
