+++ front/src/components/login_form/Login.scss
@@ -1,17 +1,16 @@
 .formLogin {
 	display: flex;
-	justify-content: space-between;
-	align-items: flex-start;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
 	width: 100%;
 	font-size: 12px;
 
+
 	&__inputs {
-		display: flex;
+		align-items: center;
 
 		&--input {
-			display: flex;
-			flex-direction: column;
-			align-items: center;
 			margin: 5px;
 
 			.email,
@@ -20,6 +19,9 @@
 				height: 35px;
 				border-radius: 5px;
 				text-align: center;
+				font-size: 18px;
+				width: 200px;
+				height: 40px;			
 			}
 
 			p {
@@ -30,23 +32,31 @@
 
 	&__btn {
 		cursor: pointer;
-		margin-top: 5px;
 		background: none;
 		border: none;
 		color: #fefefe;
 		text-decoration: none;
-		font-size: 13px;
+		font-size: 18px;
 		height: 35px;
 		font-weight: bold;
+		justify-content: center;
+		text-align: center;
+		align-items: center;
+		margin: 15px;
 	}
+
+
 
 }
 
 @media (min-width: 768px) {
 	.formLogin {
+		display: flex;
 		justify-content: flex-end;
+		flex-direction: row;
 		font-size: 18px;
 		margin-right: 100px;
+		align-items: center;
 
 		&__inputs {
 			display: flex;
@@ -56,6 +66,7 @@
 				.email,
 				.password {
 					width: 200px;
+					height: 35px;
 				}
 
 				p {
@@ -65,9 +76,11 @@
 			}
 		}
 
-		&__btn {
-			margin-top: 11px;
-			margin-left: 6px;
+		&__btn{
+			font-size: 13px;
+
 		}
+
+
 	}
 }
