+++ front/src/components/login_form/Login.tsx
@@ -88,6 +88,7 @@ export default function Login() {
 			<button type="submit" className="formLogin__btn" onClick={login}>
 				Zaloguj się
 			</button>
+
 		</form>
 	);
 }
