+++ front/src/components/theme_switcher/ThemeSwitcher.scss
@@ -1,0 +1,14 @@
+.themeSwitcher__button {
+    position: absolute;
+    top: 24px;
+    right: 10px;
+    border-width: 0px;
+    background-color: inherit;
+    font-size: 20px;
+    color: #e3edf5;
+
+    &:hover{
+        color: #bac3ca;
+    }
+}
+
